clipborad/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altgraph==0.17.3
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
EasyProcess==1.1
entrypoint2==1.1
frozenlist==1.3.3
greenlet==2.0.2
gTTS==2.3.2
idna==3.4
logutils==0.3.5
MouseInfo==0.1.3
mss==9.0.1
multidict==6.0.4
numpy==1.25.0
openai==0.27.8
pandas==2.0.3
pefile==2023.2.7
Pillow==9.5.0
plyer==2.1.0
psutil==5.9.5
PyAutoGUI==0.9.54
pycryptodome==3.18.0
PyGetWindow==0.0.9
pyinstaller==5.13.0
pyinstaller-hooks-contrib==2023.4
PyMsgBox==1.0.9
pymsteams==0.2.2
pyocr==0.8.3
pyodbc==4.0.39
pyperclip==1.8.2
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
python-dateutil==2.8.2
pytweening==1.0.7
pytz==2023.3
pywin32==306
pywin32-ctypes==0.2.2
rainbow-logging-handler==2.2.2
requests==2.31.0
screeninfo==0.8.1
six==1.16.0
SQLAlchemy==1.4.46
tqdm==4.65.0
tzdata==2023.3
urllib3==2.0.3
yarl==1.9.2
